requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
onnxruntime>=1.18.0

# LLM API
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.1

# Web interface
flask>=3.0.0

# Data processing
tqdm>=4.65.0
